+++ test/unit/utils_test.rb
@@ -42,6 +42,9 @@ class UtilsTest < ActiveSupport::TestCase
     output = klass.get_cell_data(@options.merge(headers: true), SpreadsheetArchitect)
     assert_equal [[]], output[:headers]
 
+    output = klass.get_cell_data(@options.merge(headers: false), SpreadsheetArchitect)
+    assert_equal false, output[:headers]
+
     output = klass.get_cell_data(@options.merge(column_types: nil), SpreadsheetArchitect)
     assert_nil output[:column_types]